requirements.txt
Flask==1.0.2
pyimgur==0.6.0
PySocks==1.6.8
pyTelegramBotAPI==3.6.6
python-dateutil==2.8.0
requests==2.21.0
SQLAlchemy==1.2.18
